refactor(tests): consolidate test suite — fold *_extended files, split extractors
